go gen creates graphql client
