+++ lunatrace/bsl/backend-cdk/lib/lunatrace-backend-stack.ts
@@ -357,7 +357,7 @@ export class LunatraceBackendStack extends cdk.Stack {
     taskDef.addContainer('NPMReplicator', {
       image: ingestWorkerImage,
       logging: datadogLogDriverForService('lunatrace', 'NPMReplicator'),
-      command: ['package', 'replicate'],
+      command: ['package', 'replicate', '--resume'],
       environment: {
         LUNATRACE_GRAPHQL_SERVER_URL: 'http://localhost:8080/v1/graphql',
       },
