+++ lunatrace/bsl/ingest-worker/pkg/metadata/mapper/graphql.go
@@ -12,7 +12,7 @@
 package mapper
 
 import (
-	gql2 "github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
+	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
 	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types"
 	"time"
 
@@ -23,8 +23,8 @@ import (
 var npmV types.PackageManager = types.NPM
 
 // Map converts a fetcher.PackageMetadata into the struct required by GraphQL codegen.
-func Map(p *metadata.PackageMetadata) (*gql2.Package_insert_input, error) {
-	r := &gql2.Package_insert_input{
+func Map(p *metadata.PackageMetadata) (*gql.Package_insert_input, error) {
+	r := &gql.Package_insert_input{
 		Custom_registry:       util.Ptr(""),
 		Description:           util.Ptr(p.Description),
 		Name:                  util.Ptr(p.Name),
@@ -36,10 +36,10 @@ func Map(p *metadata.PackageMetadata) (*gql2.Package_insert_input, error) {
 	return r, nil
 }
 
-func mapReleases(r []metadata.Release) *gql2.Package_release_arr_rel_insert_input {
-	m := make([]*gql2.Package_release_insert_input, len(r))
+func mapReleases(r []metadata.Release) *gql.Package_release_arr_rel_insert_input {
+	m := make([]*gql.Package_release_insert_input, len(r))
 	for i, rl := range r {
-		m[i] = &gql2.Package_release_insert_input{
+		m[i] = &gql.Package_release_insert_input{
 			Publishing_maintainer: mapMaintainer(rl.PublishingMaintainer),
 
 			Release_time:      util.Ptr(rl.ReleaseTime),
@@ -59,23 +59,23 @@ func mapReleases(r []metadata.Release) *gql2.Package_release_arr_rel_insert_inpu
 		return nil
 	}
 
-	return &gql2.Package_release_arr_rel_insert_input{
+	return &gql.Package_release_arr_rel_insert_input{
 		Data:        m,
-		On_conflict: gql2.ReleaseOnConflict,
+		On_conflict: metadata.ReleaseOnConflict,
 	}
 }
 
-func mapDependencies(ds []metadata.Dependency) *gql2.Package_release_dependency_arr_rel_insert_input {
-	m := make([]*gql2.Package_release_dependency_insert_input, len(ds))
+func mapDependencies(ds []metadata.Dependency) *gql.Package_release_dependency_arr_rel_insert_input {
+	m := make([]*gql.Package_release_dependency_insert_input, len(ds))
 	for i, dep := range ds {
-		m[i] = &gql2.Package_release_dependency_insert_input{
+		m[i] = &gql.Package_release_dependency_insert_input{
 			// create a stub entry for packages which are not yet analyzed.
-			Dependency_package: &gql2.Package_obj_rel_insert_input{
-				Data: &gql2.Package_insert_input{
+			Dependency_package: &gql.Package_obj_rel_insert_input{
+				Data: &gql.Package_insert_input{
 					Name:            util.Ptr(dep.Name),
 					Package_manager: &npmV,
 				},
-				On_conflict: gql2.PackageOnConflict,
+				On_conflict: metadata.PackageOnConflict,
 			},
 			Package_name:          util.Ptr(dep.Name),
 			Package_version_query: util.Ptr(dep.Version),
@@ -86,16 +86,16 @@ func mapDependencies(ds []metadata.Dependency) *gql2.Package_release_dependency_
 		return nil
 	}
 
-	return &gql2.Package_release_dependency_arr_rel_insert_input{
+	return &gql.Package_release_dependency_arr_rel_insert_input{
 		Data:        m,
-		On_conflict: gql2.PackageReleaseDependencyOnConflict,
+		On_conflict: metadata.PackageReleaseDependencyOnConflict,
 	}
 }
 
-func mapMaintainers(p []metadata.Maintainer) *gql2.Package_package_maintainer_arr_rel_insert_input {
-	m := make([]*gql2.Package_package_maintainer_insert_input, len(p))
+func mapMaintainers(p []metadata.Maintainer) *gql.Package_package_maintainer_arr_rel_insert_input {
+	m := make([]*gql.Package_package_maintainer_insert_input, len(p))
 	for i, pm := range p {
-		m[i] = &gql2.Package_package_maintainer_insert_input{
+		m[i] = &gql.Package_package_maintainer_insert_input{
 			Maintainer: mapMaintainer(pm),
 		}
 
@@ -105,19 +105,19 @@ func mapMaintainers(p []metadata.Maintainer) *gql2.Package_package_maintainer_ar
 		return nil
 	}
 
-	return &gql2.Package_package_maintainer_arr_rel_insert_input{
+	return &gql.Package_package_maintainer_arr_rel_insert_input{
 		Data:        m,
-		On_conflict: gql2.PackageMaintainerOnConflict,
+		On_conflict: metadata.PackageMaintainerOnConflict,
 	}
 }
 
-func mapMaintainer(pm metadata.Maintainer) *gql2.Package_maintainer_obj_rel_insert_input {
-	return &gql2.Package_maintainer_obj_rel_insert_input{
-		Data: &gql2.Package_maintainer_insert_input{
+func mapMaintainer(pm metadata.Maintainer) *gql.Package_maintainer_obj_rel_insert_input {
+	return &gql.Package_maintainer_obj_rel_insert_input{
+		Data: &gql.Package_maintainer_insert_input{
 			Email:           util.Ptr(pm.Email),
 			Name:            util.Ptr(pm.Name),
 			Package_manager: &npmV,
 		},
-		On_conflict: gql2.MaintainerOnConflict,
+		On_conflict: metadata.MaintainerOnConflict,
 	}
 }
