+++ lunatrace/bsl/ingest-worker/pkg/metadata/replicator.go
@@ -3,6 +3,7 @@ package metadata
 import "context"
 
 type Replicator interface {
+	GetLastReplicatedOffset() (int, error)
 	ReplicateSince(ctx context.Context, offset int) error
 	InitialReplication(ctx context.Context) error
 }
