+++ lunatrace/bsl/ingest-worker/pkg/vulnerability/ingester.go
@@ -15,7 +15,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/Khan/genqlient/graphql"
-	gql2 "github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
+	"github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
 	"github.com/rs/zerolog/log"
 	"github.com/schollz/progressbar/v3"
 	"github.com/vektah/gqlparser/v2/gqlerror"
@@ -79,7 +79,7 @@ func (f FileAdvisoryIngester) Ingest(ctx context.Context, source string, advisor
 
 		filenamesChunk := advisoryFiles[i:endIdx]
 
-		var vulnerabilitiesInsert []*gql2.Vulnerability_insert_input
+		var vulnerabilitiesInsert []*gql.Vulnerability_insert_input
 		vulnerabilitiesInsert, err = generateBulkVulnerabilityInsertQuery(source, filenamesChunk)
 		if err != nil {
 			log.Warn().
@@ -132,10 +132,10 @@ func collectAdvisoryFiles(isDir bool, advisoryLocation string) ([]string, error)
 
 func (f FileAdvisoryIngester) upsertVulnerabilities(
 	ctx context.Context,
-	vulnerabilitiesInsert []*gql2.Vulnerability_insert_input,
+	vulnerabilitiesInsert []*gql.Vulnerability_insert_input,
 ) ([]string, error) {
 
-	resp, err := gql2.UpsertVulnerabilities(ctx, f.GQLClient, vulnerabilitiesInsert, gql2.VulnerabilityOnConflict)
+	resp, err := gql.UpsertVulnerabilities(ctx, f.GQLClient, vulnerabilitiesInsert, vulnerabilityOnConflict)
 	if err != nil {
 		// TODO (cthompson) If there is an error that happens during an upsert, it will be difficult to determine
 		// which of the vulnerabilities caused the error. Some better error handling here would help.
