+++ lunatrace/bsl/ingest-worker/pkg/vulnerability/onconflict.go
@@ -1,0 +1,18 @@
+package vulnerability
+
+import . "github.com/lunasec-io/lunasec/lunatrace/gogen/gql"
+
+var vulnerabilityOnConflict = &Vulnerability_on_conflict{
+	Constraint: Vulnerability_constraintVulnerabilitySourceSourceIdIdx,
+	Update_columns: []Vulnerability_update_column{
+		Vulnerability_update_columnDatabaseSpecific,
+		Vulnerability_update_columnModified,
+		Vulnerability_update_columnPublished,
+		Vulnerability_update_columnWithdrawn,
+		Vulnerability_update_columnDetails,
+		Vulnerability_update_columnSummary,
+		Vulnerability_update_columnUpstreamData,
+		Vulnerability_update_columnReviewedBySource,
+		Vulnerability_update_columnSeverityName,
+	},
+}
