+++ lunatrace/gogen/generate.go
@@ -1,0 +1,4 @@
+package gogen
+
+//go:generate go run cmd/main.go
+//go:generate go run github.com/Khan/genqlient genqlient.yaml
