+++ lunatrace/gogen/genqlient.yaml
@@ -1,4 +1,4 @@
-schema: gql/schema.graphql
+schema: schema.graphql
 
 operations:
   - gql/graphql/*.graphql
@@ -42,10 +42,10 @@ bindings:
   numeric:
     type: float64
   package_manager:
-    type: github.com/lunasec-io/lunasec/lunatrace/gen/gql/types.PackageManager
+    type: github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types.PackageManager
   license_source:
-    type: github.com/lunasec-io/lunasec/lunatrace/gen/gql/types.LicenseSource
+    type: github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types.LicenseSource
   affected_range_type:
-    type: github.com/lunasec-io/lunasec/lunatrace/gen/gql/types.AffectedRangeType
+    type: github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types.AffectedRangeType
   reference_type:
-    type: github.com/lunasec-io/lunasec/lunatrace/gen/gql/types.ReferenceType
+    type: github.com/lunasec-io/lunasec/lunatrace/gogen/gql/types.ReferenceType
