--- lunatrace/gogen/introspection.go
@@ -1,97 +1,0 @@
-package gogen
-
-const introspectionQuery = `
-    query IntrospectionQuery {
-      __schema {
-        ${schemaDescription}
-        queryType { name }
-        mutationType { name }
-        subscriptionType { name }
-        types {
-          ...FullType
-        }
-        directives {
-          name
-          ${descriptions}
-          ${directiveIsRepeatable}
-          locations
-          args${inputDeprecation('(includeDeprecated: true)')} {
-            ...InputValue
-          }
-        }
-      }
-    }
-    fragment FullType on __Type {
-      kind
-      name
-      ${descriptions}
-      ${specifiedByUrl}
-      fields(includeDeprecated: true) {
-        name
-        ${descriptions}
-        args${inputDeprecation('(includeDeprecated: true)')} {
-          ...InputValue
-        }
-        type {
-          ...TypeRef
-        }
-        isDeprecated
-        deprecationReason
-      }
-      inputFields${inputDeprecation('(includeDeprecated: true)')} {
-        ...InputValue
-      }
-      interfaces {
-        ...TypeRef
-      }
-      enumValues(includeDeprecated: true) {
-        name
-        ${descriptions}
-        isDeprecated
-        deprecationReason
-      }
-      possibleTypes {
-        ...TypeRef
-      }
-    }
-    fragment InputValue on __InputValue {
-      name
-      ${descriptions}
-      type { ...TypeRef }
-      defaultValue
-      ${inputDeprecation('isDeprecated')}
-      ${inputDeprecation('deprecationReason')}
-    }
-    fragment TypeRef on __Type {
-      kind
-      name
-      ofType {
-        kind
-        name
-        ofType {
-          kind
-          name
-          ofType {
-            kind
-            name
-            ofType {
-              kind
-              name
-              ofType {
-                kind
-                name
-                ofType {
-                  kind
-                  name
-                  ofType {
-                    kind
-                    name
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  `
